using vagrant external HD instead of loop btrfs, switching to direct-lvm

+++ Vagrantfile
@@ -1,5 +1,7 @@
 # -*- mode: ruby -*-
 # vi: set ft=ruby :
+file_to_disk = './docker_disk.vdi'
+
 Vagrant.configure(2) do |config|
   config.vm.box = "Sabayon/spinbase-amd64"
   config.vm.provider "virtualbox" do |vb|
@@ -7,7 +9,31 @@ Vagrant.configure(2) do |config|
      vb.gui = false
      vb.memory = "6096"
      vb.cpus = 3
+     unless File.exist?(file_to_disk)
+      vb.customize ['createhd', '--filename', file_to_disk, '--size', 200 * 1024]
+     end
+    vb.customize ['storageattach', :id, '--storagectl', 'SATA Controller', '--port', 1, '--device', 0, '--type', 'hdd', '--medium', file_to_disk]
+
   end
+
+
+  config.vm.provision "shell", inline: <<-SHELL
+ set -e
+ set -x
+
+ if [ -f /etc/provision_env_disk_added_date ]
+ then
+    echo "Provision runtime already done."
+    exit 0
+ fi
+
+sudo pvcreate /dev/sdb
+vgcreate vg-docker /dev/sdb
+sudo lvcreate -L 190G -n data vg-docker
+sudo lvcreate -L 9G -n metadata vg-docker
+date > /etc/provision_env_disk_added_date
+   SHELL
+
   config.vm.provision "shell", inline: <<-SHELL
     mkdir -p /usr/portage/licenses/
     rsync -av -H -A -X --delete-during "rsync://rsync.at.gentoo.org/gentoo-portage/licenses/" "/usr/portage/licenses/"
@@ -21,25 +47,9 @@ Vagrant.configure(2) do |config|
     mkdir /etc/systemd/system/docker.service.d/
     echo "[Service]
 ExecStart=
-ExecStart=/usr/bin/docker daemon --storage-driver=btrfs -H fd://
+ExecStart=/usr/bin/docker daemon --storage-driver=devicemapper --storage-opt dm.datadev=/dev/vg-docker/data --storage-opt dm.metadatadev=/dev/vg-docker/metadata -H fd://
 " > /etc/systemd/system/docker.service.d/vagrant_mount.conf
     # append -g /vagrant/docker_cache/ to args to specify a default location
-    if [ ! -f "/vagrant/btrfs.img" ]; then
-      echo "Generating BTRFS image. Hang on, it could take a while. If you don't want me to create that for you, create a btrfs.img file inside the buildpsec directory"
-      echo "> equo i sys-fs/btrfs-progs"
-      echo "> dd if=/dev/zero of=btrfs.img count=<SIZE> bs=1G"
-      echo "> losetup /dev/loop0 btrfs.img"
-      echo "> mkfs.btrfs /dev/loop0"
-      dd if=/dev/zero of=/vagrant/btrfs.img count=200 bs=1G
-      losetup /dev/loop0 /vagrant/btrfs.img
-      mkfs.btrfs /dev/loop0
-    else
-      echo "BTRFS image locally detected, using it"
-      losetup /dev/loop0 /vagrant/btrfs.img
-    fi
-    mkdir /var/lib/docker
-    mount /dev/loop0 /var/lib/docker
-    echo "/vagrant/btrfs.img /var/lib/docker btrfs loop 0 0" >> /etc/fstab
 
 
     systemctl daemon-reload
@@ -54,4 +64,10 @@ ExecStart=/usr/bin/docker daemon --storage-driver=btrfs -H fd://
     timedatectl set-ntp true
     echo "@@@@ Provision finished, ensure everything is set up for deploy, suggestion is to reboot the machine to ensure docker is working correctly"
   SHELL
+
+config.vm.provision :shell, run: "always", inline: <<-SHELL
+vgscan
+vgchange -a y
+SHELL
+
 end
